requirements.txt
google-genaipillownumpy